+++ mne/epochs.py
@@ -918,7 +918,7 @@ class _BaseEpochs(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             Apply projection.
         bandwidth : float
             The bandwidth of the multi taper windowing function in Hz.
-            The default value is a window half-bandwidth of 4.
+            The default value is a window half-bandwidth of 4 Hz.
         adaptive : bool
             Use adaptive weights to combine the tapered spectra into PSD
             (slow, use n_jobs >> 1 to speed up computation).